add queue pkg for requeue interval
