+++ controllers/control-plane/kcp_suite_test.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	operatorv1beta2 "github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	"github.com/kyma-project/lifecycle-manager/pkg/queue"
 	_ "github.com/open-component-model/ocm/pkg/contexts/ocm"
 
 	"k8s.io/client-go/rest"
@@ -115,7 +116,7 @@ var _ = BeforeSuite(func() {
 		})
 	Expect(err).ToNot(HaveOccurred())
 
-	intervals := controllers.RequeueIntervals{
+	intervals := queue.RequeueIntervals{
 		Success: 1 * time.Second,
 		Busy:    100 * time.Millisecond,
 		Error:   100 * time.Millisecond,
