+++ controllers/kyma_controller/suite_test.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	"github.com/kyma-project/lifecycle-manager/pkg/queue"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	_ "github.com/open-component-model/ocm/pkg/contexts/ocm"
@@ -109,7 +110,7 @@ var _ = BeforeSuite(func() {
 		})
 	Expect(err).ToNot(HaveOccurred())
 
-	intervals := controllers.RequeueIntervals{
+	intervals := queue.RequeueIntervals{
 		Success: 1 * time.Second,
 		Busy:    100 * time.Millisecond,
 		Error:   100 * time.Millisecond,
